requirements.txt
pyserial==3.0.1
python-jenkins==0.4.12
